+++ glean-core/src/metrics/mod.rs
@@ -23,6 +23,7 @@ mod memory_distribution;
 mod memory_unit;
 mod metrics_enabled_config;
 mod numerator;
+mod object;
 mod ping;
 mod quantity;
 mod rate;
@@ -54,6 +55,7 @@ pub use self::labeled::{LabeledBoolean, LabeledCounter, LabeledMetric, LabeledSt
 pub use self::memory_distribution::MemoryDistributionMetric;
 pub use self::memory_unit::MemoryUnit;
 pub use self::numerator::NumeratorMetric;
+pub use self::object::ObjectMetric;
 pub use self::ping::PingType;
 pub use self::quantity::QuantityMetric;
 pub use self::rate::{Rate, RateMetric};
@@ -141,6 +143,8 @@ pub enum Metric {
     Url(String),
     /// A Text metric. See [`TextMetric`] for more information.
     Text(String),
+    /// An Object metric. See [`ObjectMetric`] for more information.
+    Object(String),
 }
 
 /// A [`MetricType`] describes common behavior across all metrics.
@@ -251,6 +255,7 @@ impl Metric {
             Metric::MemoryDistribution(_) => "memory_distribution",
             Metric::Jwe(_) => "jwe",
             Metric::Text(_) => "text",
+            Metric::Object(_) => "object",
         }
     }
 
@@ -280,6 +285,9 @@ impl Metric {
             Metric::MemoryDistribution(hist) => json!(memory_distribution::snapshot(hist)),
             Metric::Jwe(s) => json!(s),
             Metric::Text(s) => json!(s),
+            Metric::Object(s) => {
+                serde_json::from_str(s).expect("object storage should have been json")
+            }
         }
     }
 }